+++ RELEASENOTES.md
@@ -24,6 +24,8 @@
         `DefaultPreloadManager` and the custom
         `PreloadMediaSource.PreloadControl` implementations to preload the next
         source or take other actions.
+    *   Fix bug where silence skipping at the end of items can trigger a
+        playback exception.
 *   Transformer:
     *   Work around a decoder bug where the number of audio channels was capped
         at stereo when handling PCM input.